+++ Makefile.pre.in
@@ -2363,6 +2363,7 @@ LIBSUBDIRS=	asyncio \
 		__phello__
 TESTSUBDIRS=	idlelib/idle_test \
 		test \
+		test/test_ast \
 		test/archivetestdata \
 		test/audiodata \
 		test/certdata \
